bump platform_tools to 33.0.3
on an apple silicon the fastboot binary was incorrect architecture.

here google publishes the latest version + changelog:

https://developer.android.com/studio/releases/platform-tools

we can see that "universal binaries" were implemented in v32.0.0

tested this and it was executing fine, however the changelog
makes it seem wise to go to the latest version.

successfully flashed my comma 3 using this version
on an M1 Mac Mini on Ventura